Add arg parser.
Now using clap-rs both to parse arguments and generate help text for
whenever... Whatever hooligan can't use the application properly. Want
to make it so that the app can accept multiple files still, but haven't
done that yet. Kind of thinking I need to do that with a file/directory
swap, kind of the way I do in the C# dictionary app.